+++ src/PopDesing.Domain/Entities/Equipamento.cs
@@ -7,7 +7,7 @@ public class Equipamento
     public string Apelido { get; set; } = string.Empty;
     public DateTime DataCompra { get; set; }
     public int Potencia { get; set; }
-    public int ValorCompra { get; set; }
+    public decimal ValorCompra { get; set; }
     public int ExpectativaVida { get; set; }
     public bool Excluido { get; private set; }
     public DateTime? DataExclusao { get; private set; }
